feat: remove direct registration, add links to telegram chats
